refactor p2pk and secrets into a new Protocol class
