+++ cashu/wallet/protocols.py
@@ -1,0 +1,16 @@
+from typing import Protocol
+
+from ..core.crypto.secp import PrivateKey
+from ..core.db import Database
+
+
+class SupportsPrivateKey(Protocol):
+    private_key: PrivateKey
+
+
+class SupportsDb(Protocol):
+    db: Database
+
+
+class SupportsKeysets(Protocol):
+    keyset_id: str
